Final result(hopefully)

+++ Database/Querries/Migrations/create_printers_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS `printers` (
     SerialNumber varchar(32) NOT NULL,
-    Model INT NOT NULL,
+    Model BIGINT NOT NULL,
     Description varchar(512),
     PRIMARY KEY (SerialNumber),
     FOREIGN KEY (Model) REFERENCES printer_models(ID)
